Stop unnecessarily coping a Partition in Dialog_Rescue_Data (#759726)
The code in on_view_clicked() copy constructed a Partition object and
then in the following 3 lines only read a couple of public member
variables from the new copy.

Making a copy of the partition is unnecessary.  Change to just creating
a constant reference to the Partition instead.

(It would also be an impediment to polymorphically using Partition
objects, except for the fact that gpart doesn't recognise LUKS
signatures so will never have to create a PartitionLUKS object).

Bug 759726 - Implement Partition object polymorphism