Populate active LUKS mountpoints with /dev/mapper/NAME entry (#760080)
Populate the canonical device name, /dev/mapper/NAME, used to access the
encrypted file system into the mount points of the Partition object.
This is the equivalent of what is already done for the Volume Group name
and SWRaid Array device.

This does get displayed in the Mount Point column in the main window,
which isn't wanted.  However the data will be needed when displaying
details of the encryption mapping in the Information dialog.  Both will
be dealt with in following commits.

Bug 760080 - Implement read-only LUKS support