Prevent Mount/Umount operation being available for LUKS (#760080)
The code currently allows attempting to mount and unmount a LUKS
partition.  It is nonsense to directly try to mount and unmount a LUKS
partition and obviously doesn't work.  For read-only LUKS support there
is no need to attempt to apply this to the encrypted file system within.
Therefore prevent these operations for LUKS partitions.

Bug 760080 - Implement read-only LUKS support