Prevent incorrect no usage values warning for luks/unknown (#760080)
Create and open a LUKS mapping but don't create any file system within.

    # cryptsetup luksFormat /dev/sdb5
    # cryptsetup luksOpen /dev/sdb5 sdb5_crypt

GParted was incorrectly reporting this warning:

    Unable to read the contents of this file system!
    Because of this some operations may be unavailable.
    The cause might be a missing software package.
    The following list of software packages is required for luks file
    system support:  dmsetup.

This is even though the usage figures for the on disk LUKS encryption
are fully known.  See luks::set_used_sectors().

This was occurring because derived PartitionLUKS::set_usage_known()
was checking usage figures for the outer LUKS and the inner encrypted
file system, unknown in this case.  Correct when displaying figures in
the UI, but not correct in GParted_Core::set_used_sectors() when only
checking the outer LUKS usage figures were set correctly.  Fix this.

Bug 760080 - Implement read-only LUKS support