Display progress from the single ProgressBar in the GUI (#760709)
Change the Applying pending operations dialog so that it takes it source
of progress from the single ProgressBar object, rather than the fraction
value in every OperationDetail object.  Also remove ProgressBar
debugging now that it is being used to drive the UI.

NOTE:
This temporarily causes the existing file system specific progress bars
to not be shown because they still update via the fraction member in
each OperationDetail object, rather than the new ProgressBar.  This will
be corrected in following commits.

Bug 760709 - Add progress bars to XFS and EXT2/3/4 file system specific
             copy methods