Update ext2 fsck progress tracker to use the new ProgressBar (#760709)
Adapt the ext2 fsck progress tracker to use the new ProgressBar class.
Also make it track when the text progress bar has completely passed in
the output so that the progress bar can be stopped as well as started
when needed.

Bug 760709 - Add progress bars to XFS and EXT2/3/4 file system specific
             copy methods