Update ntfs resize progress tracker to use the new ProgressBar (#760709)
Adapt the ntfs resize progress tracker to use the new ProgressBar class.
Also make it track when the text progress indicator has passed in the
output so that the progress bar can be stopped as well as started when
needed.

Bug 760709 - Add progress bars to XFS and EXT2/3/4 file system specific
             copy methods