Display progress of ext2/3/4 file system specific copy and move operations (#760709)
Using e2image to copy a file system looks like this.  (Intermediate
progress lines which are constantly overwritten are indicated with ">").
    # e2image -ra -p /dev/sdb4 /dev/sdb5
    e2image 1.42.13 (17-May-2015)
    Scanning inodes...
>   Copying 0 / 276510 blocks (0%)
>   Copying 8845 / 276510 blocks (3%)
>   Copying 48433 / 276510 blocks (18%)
>   Copying 77135 / 276510 blocks (28%)
>   Copying 111311 / 276510 blocks (40%)
>   Copying 137039 / 276510 blocks (50%)
>   Copying 166189 / 276510 blocks (60%) 00:00:03 remaining at 108.20 MB/s
>   Copying 190285 / 276510 blocks (69%) 00:00:03 remaining at 106.19 MB/s
>   Copying 209675 / 276510 blocks (76%) 00:00:02 remaining at 102.38 MB/s
>   Copying 238219 / 276510 blocks (86%) 00:00:01 remaining at 103.39 MB/s
>   Copying 256692 / 276510 blocks (93%) 00:00:00 remaining at 100.27 MB/s
    Copied 276510 / 276510 blocks (100%) in 00:00:10 at 108.01 MB/s

Note that the copying figures are reported in file system block size
units and the progress information is written to stderr, hence needing
these two previous commits:
    Record file system block size where known (#760709)
    Call any FS specific progress trackers for stderr updates too (#760709)

Add progress tracking function for e2image command.  Also tracks when
the text progress indicator has passed in the output so that the
progress bar can be stopped as well as started when needed.

Bug 760709 - Add progress bars to XFS and EXT2/3/4 file system specific
             copy methods