Call any FS specific progress trackers for stderr updates too (#760709)
So far the signal_progress callback slot was only emitted when standard
output from the file system specific command was updated.  This was okay
as all the commands until now wrote their progress information to
stdout.  However e2image writes its progress information to stderr,
therefore also emit signal_progress when stderr is updated too.

This does mean that the file system specific *_progress() tracking
callbacks will be called when either of the OperationDetail objects
containing stdout or stderr are updated.  Therefore the trackers may be
called when there is no update to the stream from which it is parsing
the progress information.  This is not a problem as the tracker will
just update the progress bar with the same information it already has.
Also it won't happen much as only e2image is known to write to both
streams, and then only one line to stdout and the updated progress
information to stderr.  This is just an observation and not an issue
which needs coding around.

Bug 760709 - Add progress bars to XFS and EXT2/3/4 file system specific
             copy methods